Calculate AICc for a permutational multivariate analysis of variance (PERMANOVA)

#' This function calculates the Akaike's Information Criterion (AICc) for a permutational multivariate analysis of variance (PERMANOVA) model. The AICc is a modified version of the Akaike Information Criterion (AIC) that is more appropriate for small sample sizes and high-dimensional models.

AICc_permanova2(adonis2_model)

Arguments

adonis2_model

An object of class adonis2 from the vegan package

Value

A data frame with the AICc, the number of parameters (k) and the number of observations (N).

Details

The AICc calculation for a PERMANOVA model is:

$$AICc = AIC + \frac{2k(k+1)}{n-k-1}$$

where AIC is the Akaike Information Criterion, k is the number of parameters in the model (excluding the intercept), and n is the number of observations.

References

Zuur, A. F., Ieno, E. N., Walker, N. J., Saveliev, A. A., & Smith, G. M. (2009). Mixed effects models and extensions in ecology with R. Springer Science & Business Media.

See also

Examples

library(vegan)
#> Loading required package: permute
#> Loading required package: lattice
#> This is vegan 2.6-4
data(dune)
data(dune.env)

# Run PERMANOVA using adonis2

Model <- adonis2(dune ~ Management*A1, data = dune.env)

# Calculate AICc
AICc_permanova2(Model)
#>        AICc k  N
#> 1 -19.06395 8 20