Fit PERMANOVA models and arrange by AICc

This function fits PERMANOVA models for all combinations of variables in a given dataset, and arranges the models by Akaike Information Criterion (AICc) score. The function also calculates the maximum variance inflation factor (max_vif) for each model.

fit_models(
  all_forms,
  veg_data,
  env_data,
  method = "bray",
  ncores = 2,
  log = TRUE,
  logfile = "log.txt",
  multiple = 100,
  strata = NULL,
  verbose = FALSE
)

Arguments

all_forms

A data frame generated by make_models

veg_data

A dataset with vegetation presence absense or abundance data

env_data

A dataset with the variables described in all_froms

method

method for distance from vegdist

ncores

An integer specifying the number of cores to use for parallel processing

log

logical if true, a log file will be generated

logfile

the text file that will be generated as a log

multiple

after how many loops to write a log file

strata

a block variable similar to the use in adonis2

verbose

logical, defaults TRUE, sends messages about processing times

Value

A data.frame with fitted models arranged by AICc, including the formula used, the number of explanatory variables, R2, adjusted R2, and the AICc and max VIF.

References

Anderson, M. J. (2001). A new method for non-parametric multivariate analysis of variance. Austral Ecology, 26(1), 32-46. https://doi.org/10.1111/j.1442-9993.2001.01070.pp.x

Examples

# \donttest{
library(vegan)
data(dune)
data(dune.env)

AllModels <- make_models(vars = c("A1", "Moisture", "Manure"))
#> 1 of 3 ready 2023-04-06 05:33:06
#> 2 of 3 ready 2023-04-06 05:33:10
#> 3 of 3 ready 2023-04-06 05:33:12

fit_models(all_forms = AllModels,
           veg_data = dune,
           env_data = dune.env)
#>                                form      AICc  max_vif         A1  Moisture
#> 1               Distance ~ Moisture -30.36319 0.000000         NA 0.4019903
#> 2                     Distance ~ A1 -29.72347 0.000000 0.16816657        NA
#> 3                      Distance ~ 1 -28.52467 0.000000         NA        NA
#> 4          Distance ~ A1 + Moisture -28.21149 3.000000 0.04230343 0.2761272
#> 5                 Distance ~ Manure -25.21490 0.000000         NA        NA
#> 6            Distance ~ A1 + Manure -25.18745 4.000000 0.12092088        NA
#> 7      Distance ~ Moisture + Manure -22.93984 4.000000         NA 0.3005223
#> 8 Distance ~ A1 + Moisture + Manure -18.59078 4.508169 0.04145173 0.2210532
#>      Manure Model
#> 1        NA    NA
#> 2        NA    NA
#> 3        NA     0
#> 4        NA    NA
#> 5 0.3544714    NA
#> 6 0.3072258    NA
#> 7 0.2530035    NA
#> 8 0.2521518    NA
# }